Set up GND updates via OAI-PMH
See https://github.com/hbz/lobid-authorities/issues/13

RDF updates are only available via OAI-PMH (unlike MARC 21 data).

Get updates for the time passed to `ConvertUpdates`, store the
data in a file specified in the config, index that on startup.

Includes new modification of the RDF model to fix indexing issue:

In some updates, the `sameAs` property object is a literal string
instead of a resource URI. This resulted in indexing problems.

Some notes on production setup:

If update transformation was successful, the end date is stored in
a file specified in the config. This date should be passed to
`ConvertBaseline` on the next updates transformation run.

The OCLC harvester library used here completely stops if one of the
resumption token calls fails, which happens for updates on
2017-07-19.

The mechanism mentioned above would cover that if it's a temporal
issue, since it would retry the update time over and over again.
For persisting issues this is no good. Wondering how this should be
handled in OAI-PMH (current resumption call fails, we get no new
resumption token, how to skip less than a full day of updates).