Remove statements that would result in uncompacted index fields
Before this, when new fields were added in the GND source data,
they would result in non-compacted URI field names, since they are
not in the context. This is checked by `checkCompactedProperties.sh`
for the test index, and production indexing is blocked if there are
uncompacted fields. With this change, we remove the statements that
would result in such fields instead and log that as an error.